chore: Update package version to 2.1.0 and refactor import paths in App component
